Kill old condition system
